codegym + codewars updates

+++ CodeGymTasks/1.JavaSyntax/src/en/codegym/task/pro/task07/task0709/Solution.java
@@ -5,8 +5,8 @@ Very correct types
 */
 
 public class Solution {
-    long a = 109 + 15;
-    int b = a * 2;
-    short c = a / b;
-    byte d = a + b - c;
+    long a =  (109 + 15);
+    int b = (int)(a * 2);
+    short c = (short)(a / b);
+    byte d = (byte)(a + b - c);
 }
